+++ module/config/system.mjs
@@ -19,7 +19,8 @@ export const ROLL_TYPE = Object.freeze({
   DAMAGE: "damage",
   ATTACK: "attack",
   MATERIAL: "material",
-  WEAPON: "weapon"
+  WEAPON: "weapon",
+  SANITY: "sanity"
 })
 
 /**